+++ src/main/java/com/ktb/answer/service/impl/AnswerQueryServiceImpl.java
@@ -25,8 +25,8 @@ import com.ktb.answer.service.CursorCodec;
 import com.ktb.hashtag.domain.AnswerHashtag;
 import com.ktb.hashtag.repository.AnswerHashtagRepository;
 import com.ktb.interview.application.service.flow.InterviewSessionFeedbackQueryFlowService;
-import com.ktb.interview.session.dto.response.InterviewSessionFinalFeedbackResponse;
-import com.ktb.interview.session.exception.InterviewSessionInvalidStateException;
+import com.ktb.interview.dto.response.InterviewSessionFinalFeedbackResponse;
+import com.ktb.interview.exception.InterviewSessionInvalidStateException;
 import com.ktb.metric.domain.AnswerMetric;
 import com.ktb.metric.repository.AnswerMetricRepository;
 import com.ktb.question.domain.Question;
