+++ src/main/java/com/ktb/interview/application/service/flow/FinalFeedbackAnchorResolver.java
@@ -4,7 +4,7 @@ import com.ktb.answer.domain.AnswerType;
 import com.ktb.interview.domain.InterviewHistoryItem;
 import com.ktb.interview.domain.InterviewQuestionSnapshot;
 import com.ktb.interview.domain.InterviewSession;
-import com.ktb.interview.session.exception.InterviewSessionInvalidStateException;
+import com.ktb.interview.exception.InterviewSessionInvalidStateException;
 import com.ktb.question.domain.Question;
 import com.ktb.question.domain.QuestionType;
 import com.ktb.question.exception.QuestionNotFoundException;
