+++ src/main/java/com/ktb/interview/application/service/flow/InterviewPracticeSubmissionFlowService.java
@@ -6,8 +6,8 @@ import com.ktb.interview.domain.InterviewHistoryItem;
 import com.ktb.interview.domain.InterviewQuestionSnapshot;
 import com.ktb.interview.domain.InterviewSession;
 import com.ktb.interview.domain.InterviewSessionStatus;
-import com.ktb.interview.session.dto.request.PracticeAnswerSubmitRequest;
-import com.ktb.interview.session.dto.response.InterviewPracticeSubmitResponse;
+import com.ktb.interview.dto.request.PracticeAnswerSubmitRequest;
+import com.ktb.interview.dto.response.InterviewPracticeSubmitResponse;
 import com.ktb.interview.session.service.InterviewSessionService;
 import com.ktb.interview.validator.InterviewSubmissionValidator;
 import com.ktb.question.domain.Question;
