+++ src/main/java/com/ktb/interview/application/service/flow/InterviewRealSubmissionFlowService.java
@@ -10,8 +10,8 @@ import com.ktb.interview.config.InterviewSessionProperties;
 import com.ktb.interview.domain.InterviewQuestionSnapshot;
 import com.ktb.interview.domain.InterviewSession;
 import com.ktb.interview.domain.InterviewSessionStatus;
-import com.ktb.interview.session.dto.request.RealAnswerSubmitRequest;
-import com.ktb.interview.session.dto.response.InterviewRealSubmitResponse;
+import com.ktb.interview.dto.request.RealAnswerSubmitRequest;
+import com.ktb.interview.dto.response.InterviewRealSubmitResponse;
 import com.ktb.interview.session.service.InterviewSessionService;
 import com.ktb.interview.validator.InterviewSubmissionValidator;
 import com.ktb.question.domain.QuestionCategory;
