+++ src/main/java/com/ktb/interview/application/service/flow/InterviewSessionCreateFlowService.java
@@ -4,10 +4,10 @@ import com.ktb.answer.domain.AnswerType;
 import com.ktb.answer.domain.TurnType;
 import com.ktb.interview.domain.InterviewQuestionSnapshot;
 import com.ktb.interview.domain.InterviewSession;
-import com.ktb.interview.session.dto.request.InterviewSessionCreateRequest;
-import com.ktb.interview.session.dto.response.InterviewSessionCreateResponse;
-import com.ktb.interview.session.exception.InterviewSessionInvalidInputException;
-import com.ktb.interview.session.exception.InterviewSessionInvalidStateException;
+import com.ktb.interview.dto.request.InterviewSessionCreateRequest;
+import com.ktb.interview.dto.response.InterviewSessionCreateResponse;
+import com.ktb.interview.exception.InterviewSessionInvalidInputException;
+import com.ktb.interview.exception.InterviewSessionInvalidStateException;
 import com.ktb.interview.session.service.InterviewSessionService;
 import com.ktb.question.domain.Question;
 import com.ktb.question.domain.QuestionCategory;
