+++ src/main/java/com/ktb/interview/application/service/flow/InterviewSessionFinalFeedbackFlowService.java
@@ -10,8 +10,8 @@ import com.ktb.interview.domain.InterviewHistoryItem;
 import com.ktb.interview.domain.InterviewSession;
 import com.ktb.interview.domain.InterviewSessionFeedback;
 import com.ktb.interview.domain.InterviewSessionStatus;
-import com.ktb.interview.session.dto.response.InterviewSessionFinalFeedbackResponse;
-import com.ktb.interview.session.exception.InterviewSessionInvalidStateException;
+import com.ktb.interview.dto.response.InterviewSessionFinalFeedbackResponse;
+import com.ktb.interview.exception.InterviewSessionInvalidStateException;
 import com.ktb.interview.session.mapper.InterviewSessionFeedbackMapper;
 import com.ktb.interview.session.persistence.InterviewFinalFeedbackPersistenceService;
 import com.ktb.interview.session.repository.InterviewSessionFeedbackRepository;
