+++ src/main/java/com/ktb/interview/assembler/InterviewRealHistoryAssembler.java
@@ -5,8 +5,8 @@ import com.ktb.interview.dto.ai.InterviewHistoryRequest;
 import com.ktb.interview.domain.InterviewHistoryItem;
 import com.ktb.interview.domain.InterviewQuestionSnapshot;
 import com.ktb.interview.domain.InterviewSession;
-import com.ktb.interview.session.exception.InterviewSessionInvalidInputException;
-import com.ktb.interview.session.exception.InterviewSessionInvalidStateException;
+import com.ktb.interview.exception.InterviewSessionInvalidInputException;
+import com.ktb.interview.exception.InterviewSessionInvalidStateException;
 import com.ktb.question.domain.QuestionCategory;
 import java.util.ArrayList;
 import java.util.HashMap;
