+++ src/main/java/com/ktb/interview/controller/InterviewSessionController.java
@@ -1,21 +1,21 @@
 package com.ktb.interview.controller;
 
 import com.ktb.interview.dto.ai.InterviewFeedbackDataResponse;
-import com.ktb.interview.session.dto.request.InterviewSessionCreateRequest;
-import com.ktb.interview.session.dto.request.PracticeAnswerSubmitRequest;
-import com.ktb.interview.session.dto.request.RealAnswerSubmitRequest;
-import com.ktb.interview.session.dto.request.SessionFinalFeedbackRequest;
-import com.ktb.interview.session.dto.response.InterviewPracticeSubmitResponse;
-import com.ktb.interview.session.dto.response.InterviewSessionCreateResponse;
-import com.ktb.interview.session.dto.response.InterviewSessionFinalFeedbackResponse;
-import com.ktb.interview.session.dto.response.InterviewRealSubmitResponse;
-import com.ktb.interview.session.dto.response.InterviewSessionStateResponse;
-import com.ktb.interview.session.dto.response.SessionFeedbackFailedResponse;
-import com.ktb.interview.session.dto.response.SessionFeedbackPendingResponse;
+import com.ktb.interview.dto.request.InterviewSessionCreateRequest;
+import com.ktb.interview.dto.request.PracticeAnswerSubmitRequest;
+import com.ktb.interview.dto.request.RealAnswerSubmitRequest;
+import com.ktb.interview.dto.request.SessionFinalFeedbackRequest;
+import com.ktb.interview.dto.response.InterviewPracticeSubmitResponse;
+import com.ktb.interview.dto.response.InterviewSessionCreateResponse;
+import com.ktb.interview.dto.response.InterviewSessionFinalFeedbackResponse;
+import com.ktb.interview.dto.response.InterviewRealSubmitResponse;
+import com.ktb.interview.dto.response.InterviewSessionStateResponse;
+import com.ktb.interview.dto.response.SessionFeedbackFailedResponse;
+import com.ktb.interview.dto.response.SessionFeedbackPendingResponse;
 import com.ktb.interview.application.InterviewSessionManagementService;
 import com.ktb.interview.application.InterviewSubmissionService;
 import com.ktb.interview.domain.InterviewSessionStatus;
-import com.ktb.interview.session.exception.InterviewSessionInvalidStateException;
+import com.ktb.interview.exception.InterviewSessionInvalidStateException;
 import com.ktb.auth.security.adapter.SecurityUserAccount;
 import com.ktb.common.dto.ApiResponse;
 import com.ktb.common.util.HttpRequestUtils;
