+++ src/main/java/com/ktb/interview/dto/request/RealAnswerSubmitRequest.java
@@ -1,4 +1,4 @@
-package com.ktb.interview.session.dto.request;
+package com.ktb.interview.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonAlias;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
