+++ src/main/java/com/ktb/interview/dto/request/SessionFinalFeedbackRequest.java
@@ -1,4 +1,4 @@
-package com.ktb.interview.session.dto.request;
+package com.ktb.interview.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
