+++ src/main/java/com/ktb/interview/mapper/InterviewSubmissionResponseMapper.java
@@ -10,15 +10,15 @@ import com.ktb.interview.dto.ai.InterviewTopicFeedbackResponse;
 import com.ktb.interview.domain.InterviewHistoryItem;
 import com.ktb.interview.domain.InterviewQuestionSnapshot;
 import com.ktb.interview.domain.InterviewSessionStatus;
-import com.ktb.interview.session.dto.response.InterviewFinalFeedbackMetricResponse;
-import com.ktb.interview.session.dto.response.InterviewHistoryResponse;
-import com.ktb.interview.session.dto.response.InterviewNextQuestionTurnResponse;
-import com.ktb.interview.session.dto.response.InterviewRealSubmitResponse;
-import com.ktb.interview.session.dto.response.InterviewSessionBadCaseFeedbackResponse;
-import com.ktb.interview.session.dto.response.InterviewSessionFinalFeedbackResponse;
-import com.ktb.interview.session.dto.response.InterviewSessionKeywordResultResponse;
-import com.ktb.interview.session.dto.response.InterviewSessionOverallFeedbackResponse;
-import com.ktb.interview.session.dto.response.InterviewSessionTopicFeedbackResponse;
+import com.ktb.interview.dto.response.InterviewFinalFeedbackMetricResponse;
+import com.ktb.interview.dto.response.InterviewHistoryResponse;
+import com.ktb.interview.dto.response.InterviewNextQuestionTurnResponse;
+import com.ktb.interview.dto.response.InterviewRealSubmitResponse;
+import com.ktb.interview.dto.response.InterviewSessionBadCaseFeedbackResponse;
+import com.ktb.interview.dto.response.InterviewSessionFinalFeedbackResponse;
+import com.ktb.interview.dto.response.InterviewSessionKeywordResultResponse;
+import com.ktb.interview.dto.response.InterviewSessionOverallFeedbackResponse;
+import com.ktb.interview.dto.response.InterviewSessionTopicFeedbackResponse;
 import java.util.List;
 import org.springframework.stereotype.Component;
