+++ src/main/java/com/ktb/interview/session/service/impl/InterviewSessionServiceImpl.java
@@ -4,9 +4,9 @@ import com.ktb.answer.domain.AnswerType;
 import com.ktb.interview.session.InterviewSessionIdProvider;
 import com.ktb.interview.config.InterviewSessionProperties;
 import com.ktb.interview.domain.InterviewSession;
-import com.ktb.interview.session.exception.InterviewSessionAccessDeniedException;
-import com.ktb.interview.session.exception.InterviewSessionExpiredException;
-import com.ktb.interview.session.exception.InterviewSessionNotFoundException;
+import com.ktb.interview.exception.InterviewSessionAccessDeniedException;
+import com.ktb.interview.exception.InterviewSessionExpiredException;
+import com.ktb.interview.exception.InterviewSessionNotFoundException;
 import com.ktb.interview.session.repository.InterviewSessionFeedbackRepository;
 import com.ktb.interview.session.repository.InterviewSessionRepository;
 import com.ktb.interview.session.service.InterviewSessionService;
