+++ src/main/java/com/ktb/interview/validator/InterviewSubmissionValidator.java
@@ -5,9 +5,9 @@ import com.ktb.answer.domain.TurnType;
 import com.ktb.interview.domain.InterviewQuestionSnapshot;
 import com.ktb.interview.domain.InterviewSession;
 import com.ktb.interview.domain.InterviewSessionStatus;
-import com.ktb.interview.session.dto.request.RealAnswerSubmitRequest;
-import com.ktb.interview.session.exception.InterviewSessionInvalidInputException;
-import com.ktb.interview.session.exception.InterviewSessionInvalidStateException;
+import com.ktb.interview.dto.request.RealAnswerSubmitRequest;
+import com.ktb.interview.exception.InterviewSessionInvalidInputException;
+import com.ktb.interview.exception.InterviewSessionInvalidStateException;
 import com.ktb.question.domain.QuestionType;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
